+++ R/tsibble-package.R
@@ -85,7 +85,7 @@
 #' @importFrom dplyr group_by ungroup group_data grouped_df group_vars
 #' @importFrom dplyr group_rows groups new_grouped_df is_grouped_df
 #' @importFrom dplyr left_join right_join full_join inner_join anti_join
-#' @importFrom dplyr bind_rows bind_cols combine with_order pull first
+#' @importFrom dplyr bind_rows bind_cols combine with_order pull first reframe
 #' @importFrom dplyr dplyr_row_slice dplyr_col_modify dplyr_reconstruct
 #' @importFrom tibble new_tibble trunc_mat
 #' @import rlang tidyselect vctrs ellipsis
@@ -104,3 +104,13 @@
 #'   key = group
 #' )
 "_PACKAGE"
+
+## This function is never called
+## Exists to suppress this NOTE:
+## "Namespaces in Imports field not imported from:"
+## https://github.com/opencpu/opencpu/blob/10469ee3ddde0d0dca85bd96d2873869d1a64cd6/R/utils.R#L156-L165
+stub <- function() {
+  ## I have to use curl directly somewhere, if I import it.
+  ## I have to import it if I want to state a minimum version.
+  if (FALSE) lifecycle::deprecate_soft()
+}